[ALPS04410870] Audio: MT6660 fix coverity
1. modify chip_rev u8 -> u16

MTK-Commit-Id: 17339343ee0781bef0069b54b267a145c57c7336

CR-Id: ALPS04410870
Feature: [Module]Audio OpenDSP
(cherry picked from commit 34af10987ea11c4ebb40b58fbf8fac472a06bc8c)

+++ sound/soc/codecs/mt6660.h
@@ -27,7 +27,7 @@ struct mt6660_chip {
 	struct platform_device *param_dev;
 	struct rt_regmap_device *regmap;
 	struct mutex var_lock;
-	u8 chip_rev;
+	u16 chip_rev;
 	u8 dev_cnt;
 	int pwr_cnt;
 };
